+++ Examples/FTMobileSDKUnitTests/Helper/FTNetworkMock.m
@@ -10,17 +10,10 @@
 #import "FTJSONUtil.h"
 typedef void (^CompletionHandler)(void);
 
-static NSString *urlStr;
 @implementation FTNetworkMock
-+ (void)registerUrlString:(NSString *)urlString{
-    urlStr = urlString;
-}
 
 + (id<OHHTTPStubsDescriptor>)networkOHHTTPStubs{
     return [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
-        if(urlStr&&urlStr.length>0){
-            return [request.URL.absoluteString isEqualToString:urlStr];
-        }
         return YES;
     } withStubResponse:^OHHTTPStubsResponse*(NSURLRequest *request) {
         NSString *data  =[FTJSONUtil convertToJsonData:@{@"data":@"Hello World!",@"code":@200}];
@@ -29,6 +22,9 @@ static NSString *urlStr;
     }];
 }
 + (id<OHHTTPStubsDescriptor>)networkOHHTTPStubsHandler:(void (^)(void))handler{
+    return [self networkOHHTTPStubsWithUrl:nil handler:handler];
+}
++ (id<OHHTTPStubsDescriptor>)networkOHHTTPStubsWithUrl:(NSString *)urlStr handler:(void (^)(void))handler{
     id<OHHTTPStubsDescriptor> stubs = [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
         if(urlStr&&urlStr.length>0){
             return [request.URL.absoluteString isEqualToString:urlStr];
@@ -37,10 +33,10 @@ static NSString *urlStr;
     } withStubResponse:^OHHTTPStubsResponse*(NSURLRequest *request) {
         NSString *data  =[FTJSONUtil convertToJsonData:@{@"data":@"Hello World!",@"code":@200}];
         NSData *requestData = [data dataUsingEncoding:NSUTF8StringEncoding];
-            if(handler){
-                dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
-                    handler();
-                });
+        if(handler){
+            dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
+                handler();
+            });
         }
         return [OHHTTPStubsResponse responseWithData:requestData statusCode:200 headers:nil];
     }];