+++ Examples/FTMobileSDKUnitTests/Resource/FTResourceInstrumentTest.m
@@ -71,14 +71,12 @@
     [FTMobileAgent startWithConfigOptions:config];
     [[FTMobileAgent sharedInstance] startRumWithConfigOptions:rumConfig];
     [[FTTrackerEventDBTool sharedManger] deleteAllDatas];
-    [FTNetworkMock registerUrlString:urlStr];
 }
 
 - (void)tearDown {
     // Put teardown code here. This method is called after the invocation of each test method in the class.
     [FTMobileAgent shutDown];
     [OHHTTPStubs removeAllStubs];
-    [FTNetworkMock registerUrlString:@""];
 }
 /** Tests that creating a shared session returns a non-nil object. */
 - (void)testSharedSession {
@@ -101,7 +99,7 @@
 - (void)testSessionWithDefaultSessionConfiguration {
     __block NSURLSessionDataTask *dataTask;
     XCTestExpectation *expectation= [self expectationWithDescription:@"Async operation timeout"];
-    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsHandler:^{
+    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsWithUrl:self.url.absoluteString handler:^{
         XCTAssertNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
         [expectation fulfill];
         [OHHTTPStubs removeStub:stubs];
@@ -120,7 +118,7 @@
 - (void)testSessionWithEphemeralSessionConfiguration {
     __block NSURLSessionDataTask *dataTask;
     XCTestExpectation *expectation= [self expectationWithDescription:@"Async operation timeout"];
-    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsHandler:^{
+    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsWithUrl:self.url.absoluteString handler:^{
         XCTAssertNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
         [expectation fulfill];
         [OHHTTPStubs removeStub:stubs];
@@ -242,11 +240,6 @@
 /** Tests using a nil delegate still results in tracking responses. */
 - (void)testSessionWithConfigurationDelegateDelegateQueueWithNilDelegate {
     [FTNetworkMock networkOHHTTPStubs];
-    __weak typeof(self) weakSelf = self;
-    __block NSURLSessionDataTask *dataTask;
-//    [FTNetworkMock registerBeforeHandler:^{
-//        XCTAssertNotNil([weakSelf getTraceHandler:dataTask]);
-//    }];
     NSURLSessionConfiguration *configuration =
     [NSURLSessionConfiguration defaultSessionConfiguration];
     NSURLSession *session = [NSURLSession sessionWithConfiguration:configuration
@@ -254,7 +247,7 @@
                                                      delegateQueue:nil];
     NSURLRequest *request = [NSURLRequest requestWithURL:self.url];
     @autoreleasepool {
-        dataTask = [session dataTaskWithRequest:request];
+        NSURLSessionDataTask *dataTask = [session dataTaskWithRequest:request];
         XCTAssertNotNil(dataTask);
         [dataTask resume];
         XCTAssertNotNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
@@ -266,7 +259,7 @@
 - (void)testDelegateClassOnlyRegisteredOnce {
     __block NSURLSessionDataTask *dataTask;
     XCTestExpectation *expectation= [self expectationWithDescription:@"Async operation timeout"];
-    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsHandler:^{
+    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsWithUrl:self.url.absoluteString handler:^{
         XCTAssertNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
         [expectation fulfill];
         [OHHTTPStubs removeStub:stubs];
@@ -326,7 +319,7 @@
 - (void)testDelegateUnimplementedURLSessionTaskDidCompleteWithError {
     __block NSURLSessionDataTask *dataTask;
     XCTestExpectation *expectation= [self expectationWithDescription:@"Async operation timeout"];
-    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsHandler:^{
+    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsWithUrl:self.url.absoluteString handler:^{
         XCTAssertNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
         [expectation fulfill];
         [OHHTTPStubs removeStub:stubs];
@@ -348,7 +341,7 @@
 - (void)testDelegateUnimplementedURLSessionTaskDidFinishCollectingMetrics {
     __block NSURLSessionDataTask *dataTask;
     XCTestExpectation *expectation= [self expectationWithDescription:@"Async operation timeout"];
-    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsHandler:^{
+    __block id<OHHTTPStubsDescriptor> stubs = [FTNetworkMock networkOHHTTPStubsWithUrl:self.url.absoluteString handler:^{
         XCTAssertNil([[FTURLSessionInterceptor shared] getTraceHandler:dataTask]);
         [expectation fulfill];
         [OHHTTPStubs removeStub:stubs];